Rename the module to not use the _
